+++ bcdi/utils/parameters.py
@@ -413,6 +413,7 @@ class PostprocessingChecker(ConfigChecker):
             self._checked_params["is_orthogonal"] = False
         else:
             self._checked_params["is_orthogonal"] = True
+            self._checked_params["is_orthogonal"] = True
 
         if (
             self._checked_params["data_frame"] == "crystal"
@@ -486,8 +487,6 @@ def valid_param(key: str, value: Any) -> Tuple[Any, bool]:
         "use_rawdata",
     }:
         valid.valid_item(value, allowed_types=bool, name=key)
-    elif key == "absorption":
-        valid.valid_item(value, allowed_types=Real, min_excluded=0, name=key)
     elif key == "actuators":
         valid.valid_container(value, container_types=dict, allow_none=True, name=key)
     elif key == "apodization_alpha":